[REFACTOR] Add a utility function to redurect users home if they are don't have an ethaccount
