+++ src/views/SubmitProposal/SubmitProposal.js
@@ -39,6 +39,7 @@ import { emojiExists } from '../../modules/twitterDb';
 import IpfsUpload from '../../components/IpfsUpload';
 // import Box from '3box';
 import ProposalCard from '../../components/ProposalCard/ProposalCard';
+import { useRedirectHomeIfNoEthAccount } from '../../utils/useCommonUtils';
 
 const BN = require('bn.js');
 
@@ -191,11 +192,7 @@ const SubmitProposal = props => {
     }
   }, 3000);
 
-  useEffect(() => {
-    if (web3Connect.loaded && !web3Connect.connected) {
-      router.history.push('/');
-    }
-  }, [web3Connect, router.history]);
+  useRedirectHomeIfNoEthAccount();
 
   const { register, handleSubmit, watch, errors } = useForm();
